start reorganising strategy query
