+++ src/store/StoreProvider.tsx
@@ -35,11 +35,6 @@ import {
   useFiatCurrencyStore,
 } from 'store/useFiatCurrencyStore';
 import {
-  defaultStrategiesStore,
-  StrategiesStore,
-  useStrategiesStore,
-} from './useStrategiesStore';
-import {
   defaultOrderBookSettingsStore,
   OrderBookSettingsStore,
   useOrderBookSettingsStore,
@@ -54,7 +49,6 @@ interface StoreContext {
   setCountryBlocked: (value: boolean | null) => void;
   sdk: SDKStore;
   tokens: TokensStore;
-  strategies: StrategiesStore;
   notifications: NotificationsStore;
   modals: ModalStore;
   trade: {
@@ -75,7 +69,6 @@ const defaultValue: StoreContext = {
   setCountryBlocked: () => {},
   sdk: defaultSDKStore,
   tokens: defaultTokensStore,
-  strategies: defaultStrategiesStore,
   notifications: defaultNotificationsStore,
   modals: defaultModalStore,
   trade: {
@@ -114,14 +107,12 @@ export const StoreProvider: FC<{ children: ReactNode }> = ({ children }) => {
   const modals = useModalStore();
   const tokens = useTokensStore();
   const fiatCurrency = useFiatCurrencyStore();
-  const strategies = useStrategiesStore();
 
   const value: StoreContext = {
     isCountryBlocked: countryBlocked,
     setCountryBlocked,
     sdk,
     tokens,
-    strategies,
     notifications,
     modals,
     trade: {
